Fixed glfwGetCurrentContext() to return null when there is no context.
As it was, it was taking the NULL pointer and wrapping it up in a "window" object, so it looked like there was always some context in place.  We need to activate a context, draw, and deactivate the context (as shown in the example) to keep everything thread safe in the Dart vm.  It helps to know when the context is active and when it's not.