Add HANDLE_INVOKE for Dart_InvokeClosure calls
This allows these calls to explicitly ignore the return value of
Dart_InvokeClosure when running in release mode. This may improve
performance, at the cost of not catching error returns from these
calls early. This is needed for builds with -Wunused-result turned
on.

Bindings were regenerated from a recent GLFW version. These are new
manual bindings needed, but these are not provided in this change.